Fix symbol hashing and provide more useful symbol repr
Hash on the value since there are multiple levels of indirection to the
underlying object.  Otherwise hashing keys never match up multiple
instances of a Python object that point to the same policy object.

Also add a repr function that will make debugging doable when dealing
with Symbol objects instead of string representations.