Document that build_ext needs to be run before build
When installing setools with python setup.py install,
setools/policyrep/qpol.py is not copied to the destination directory.
This is because the file is generated in step build_ext and Python files
are copied beforehand (in step build_py).

A simple workaround consists in running "setup.py build_ext" before
building and installing setools. Document this in the README.

Closes: https://github.com/TresysTechnology/setools/issues/173