policyrep: Clean up duplicate pointer attributes.
All PolicyObjects have a attribute "key" which is a uintptr_t equivalent
to the C pointer to the object's policy db struct.  The few places
where loading attributes was done lazily had a duplicate attribute
of the pointer type.  Instead, recast the key attribute back to the
appropriate pointer type.