setup.py: drop path prefix from man install
If the directory is a relative path, it is interpreted relative to the
installation prefix (Python’s sys.prefix for pure-Python packages,
sys.exec_prefix for packages that contain extension modules).
So the sys.prefix should not be joined because the correct one will be
added regardless. The reason why a relative path is going to work is
because the installation path is built internally joining python root
with python prefix, thus forming the correct path $(TARGET_DIR)/usr.
(https://docs.python.org/3/distutils/setupscript.html#installing-additional-files)

For this man file install scenario, joining the sys.prefix makes the
path absolute (this was previously working by accident).

Signed-off-by: Matthew Weber <matthew.weber@rockwellcollins.com>