AVRule: Change derive_expanded() permission set check to disjoint check.
Jim Carter first figured out this bug in PR #33

In class AVRUle, when deriving an expanded rule remove the check for
permissions beings a subset of the original rule.

This is not a condition that holds when attributes are expanded.

Ex/
attribute a1;
attribute a2;
type t1;
typeatrribute t1 a1, a2;
allow a1 self: c1 p1;
allow a2 self: c1 p2;

In this case the expanded rule "allow t1 t1: c1 {p1 p2};" has permissions
that are not a subset of either the allow rule with a1 or the one with a2.

Change the check to a set-disjoint check to make sure there is some
intersection between the original rule permission set and the derived
permission set.

Signed-off-by: Chris PeBenito <pebenito@ieee.org>