tox.ini: Add missing parameters for checks.
apol and sechecker were missing from all of the checks.  Also change the
testenvs to use the main dependency list so there are no errors in stating
the NetworkX and Cython dependencies.

This also changes the environment list.  It adds pep8 and lint to the
default checking and changes the python versions to a generic Python 3 so
the baseline installed Python is used.  .travis.yml will be the
only place to specify which versions of Python to test for CI purposes.

Signed-off-by: Chris PeBenito <pebenito@ieee.org>