setup.py: Reduce aggressiveness of compiler flags.
The C code is generated by Cython, so SETools is dependent on Cython's
function..  If there is some misalignment between Cython and the Python
headers, for example, the code uses functions in Python 3.8 that are
deprecated in Python 3.9, that is temporary and not SETools' bug.

The unit tests are the standard for finding issues in SETools.

This addresses the current issue with Python 3.9 compiles:

setools/policyrep.c: In function ‘__pyx_pf_7setools_9policyrep_6AVRule_9statement’:
setools/policyrep.c:92274:3: error: ‘_PyUnicode_get_wstr_length’ is deprecated [-Werror=deprecated-declarations]
92274 |   __pyx_t_1 = (__pyx_v_self->__pyx_base.rule_string != Py_None)&&(__Pyx_PyUnicode_IS_TRUE(__pyx_v_self->__pyx_base.rule_string) != 0);
      |   ^~~~~~~~~
In file included from /usr/include/python3.9/unicodeobject.h:1026,
                 from /usr/include/python3.9/Python.h:97,
                 from setools/policyrep.c:52:

Closes #54

Signed-off-by: Chris PeBenito <pebenito@ieee.org>