policyrep: move transition to separate structure in avtab
Due to change to the internal avtab structures in libsepol required for
moving transition to a separate structure, same changes need to be
applied also in the setools bindings.

Therefore, this patch adds a new structure to the avtab_datum and moves
the otype of the transition to this structure. In the next patch, this
structure will also hold filename transitions for the combination of
stype, ttype and tclass.

Signed-off-by: Juraj Marcin <juraj@jurajmarcin.com>