policyrep: add prefix/suffix matching to filename type transitions
Currently, filename transitions are stored separately from other type
enforcement rules and only support exact name matching. However, in
practice, the names contain variable parts. This leads to many
duplicated rules in the policy that differ only in the part of the name,
or it is even impossible to cover all possible combinations.

This patch reflects changes in libsepol implemented in this patch [1].

The patch adds additional filename transition tables to policydb
structure for prefix and suffix rules and updates the functions that
access them.

[1]: https://lore.kernel.org/selinux/20231108103345.4014148-1-juraj@jurajmarcin.com/

Reviewed-by: Ondrej Mosnacek <omosnace@redhat.com>
Signed-off-by: Juraj Marcin <juraj@jurajmarcin.com>