InfoFlowAnalysis/DomainTransitionAnalysis: Change edge methods to BFS.
Revise infoflow in/out and transitions in/out to use the breadth-first-
search algorithm instead of in/out edges functions.  Revise the limit for
AllPaths steps to general depth limit set to 1 step.  This allows the
default in/out to return the expected depth of 1; all paths will presumably
set the depth limit in all cases anyway.  Then callers will be able to
generate a infoflow/dta tree using a depth_limit > 1.

Signed-off-by: Chris PeBenito <pebenito@ieee.org>